normalizer-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
google-generativeai==0.3.1
requests==2.31.0
prometheus-client==0.19.0
